Narrow TradeLogRecord::priceTicks to int32_t
Matches PriceTicks, which is int32_t; the int64_t field was wider than the value it ever stores.

+++ cpp/include/infra/binary_trade_logger.hpp
@@ -5,11 +5,10 @@
 #include <string>
 #include "infra/trade_logger.hpp"
 
-// TODO: make priceTicks int32_t
 struct TradeLogRecord {
     uint64_t tradeID;
     uint64_t timestamp;
-    int64_t priceTicks;
+    int32_t priceTicks;
     uint32_t takerOrderID;
     uint32_t makerOrderID;
     int32_t qty;
